# Conflicts:
#	gradle.properties
#	src/main/java/julianh06/wynnextras/core/loader/CommandLoader.java
#	src/main/resources/fabric.mod.json

+++ src/main/java/julianh06/wynnextras/features/raid/RaidListFilter.java
@@ -60,7 +60,7 @@ public class RaidListFilter extends EasyTextInput {
     @Override
     public void drawWithTexture(DrawContext context, Identifier texture) {
         long now = System.currentTimeMillis();
-        RenderUtils.drawTexturedRect(context.getMatrices(), texture, x, y, 0.0f, width, height, width, height);
+        RenderUtils.drawTexturedRect(context.getMatrices(), texture, x, y, 0.0f, width, height, (int) width, (int) height);
         if(input.isEmpty() && !isActive) {
             context.drawTextWithShadow(MinecraftClient.getInstance().textRenderer, searchText, x + 4, y + 3, CustomColor.fromHexString("FFFFFF").asInt());
         } else {
